# back-end/docs/openapi.yaml
openapi: 3.0.0
info:
  title: UnderSounds API Authentication
  version: "1.0.0"
  description: "API for user management: registration, login, update, logout, OAuth2.0 with Google, password recovery, and refresh token via cookies."
servers:
  - url: "{protocol}://{host}/api"
    variables:
      protocol:
        default: "http"
        enum:
          - "http"
          - "https"
      host:
        default: "192.168.1.70:5000"
paths:
...
  /albums:
    get:
      summary: Get all albums.
      responses:
        "200":
          description: Album list retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
        "500":
          description: Server internal error.
    post:
      summary: Creates new album.
      requestBody:
        description: Data for creating an album.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumInput'
      responses:
        "201":
          description: Album created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        "400":
          description: Invalid data.
        "500":
          description: Server internal error.
  /albums/{id}:
    get:
      summary: Get an album by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Album obtained successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        "404":
          description: Album not found.
        "500":
          description: Server internal error.
    put:
      summary: Updates an existing album.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        description: Album data updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumInput'
      responses:
        "200":
          description: Album updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        "404":
          description: Album not found.
        "500":
          description: Server internal error.
    delete:
      summary: Deletes an album.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Album deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Album deleted successfully"
        "404":
          description: Album not found.
        "500":
          description: Server internal error.
...
  /albums/{id}/download:
    get:
      summary: Downloads a specific track from an album.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: trackId
          required: true
          schema:
            type: string
          description: ID of the track to download.
        - in: query
          name: format
          required: false
          schema:
            type: string
            enum: [mp3, wav, flac]
            default: mp3
          description: Audio format for the download.
      responses:
        "200":
          description: Audio file downloaded successfully.
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/wav:
              schema:
                type: string
                format: binary
            audio/flac:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid parameters.
        "404":
          description: Album or track not found.
        "500":
          description: Server internal error.
  /albums/{id}/download-album:
    get:
      summary: Download a complete album as a ZIP file.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: format
          required: false
          schema:
            type: string
            enum: [mp3, wav, flac]
            default: mp3
          description: Audio format for the album tracks.
      responses:
        "200":
          description: ZIP file downloaded successfully.
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid parameters.
        "404":
          description: Album not found or without tracks.
        "500":
          description: Server internal error.
...
components:
  schemas:
...
    Album:
      type: object
      properties:
        _id:
          type: string
          example: "60d0fe4f5311236168a109cb"
        title:
          type: string
          example: "Álbum Debut"
        artistId:
          type: string
          example: "60d0fe4f5311236168a109cd"
        releaseYear:
          type: number
          example: 2023
        coverImage:
          type: string
          example: "/assets/images/album1.jpg"
        genre:
          type: string
          example: "Rock"
        price:
          type: number
          example: 9.99
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
        ratings:
          type: array
          items:
            $ref: '#/components/schemas/Rating'
        vinyl:
          type: boolean
          example: true
        cd:
          type: boolean
          example: true
        cassettes:
          type: boolean
          example: false
        destacado:
          type: boolean
          example: false
    AlbumInput:
      type: object
      properties:
        title:
          type: string
          example: "Álbum Debut"
        artistId:
          type: string
          example: "60d0fe4f5311236168a109cd"
        releaseYear:
          type: number
          example: 2023
        coverImage:
          type: string
          example: "/assets/images/album1.jpg"
        genre:
          type: string
          example: "Rock"
        price:
          type: number
          example: 9.99
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/TrackInput'
        vinyl:
          type: boolean
          example: true
        cd:
          type: boolean
          example: true
        cassettes:
          type: boolean
          example: false
        destacado:
          type: boolean
          example: false
      required:
        - title
        - artistId
        - releaseYear
        - genre
        - price
    Track:
      type: object
      properties:
        id:
          type: number
          example: 1
        title:
          type: string
          example: "Canción 1"
        duration:
          type: string
          example: "3:45"
        url:
          type: string
          example: "../music/song1.mp3"
        autor:
          type: string
          example: "Nombre del Artista"
        n_reproducciones:
          type: number
          example: 1500
    TrackInput:
      type: object
      properties:
        title:
          type: string
          example: "Canción 1"
        duration:
          type: string
          example: "3:45"
        url:
          type: string
          example: "../music/song1.mp3"
        autor:
          type: string
          example: "Nombre del Artista"
      required:
        - title
        - url
    Rating:
      type: object
      properties:
        userId:
          type: string
          example: "60d0fe4f5311236168a109ca"
        rating:
          type: number
          example: 4.5
        comment:
          type: string
          example: "Gran álbum, excelentes pistas"
        profileImage:
          type: string
          example: "/assets/images/avatar1.jpg"
...